refactor(proyecto): se agrega archivo logo a assets.Se elimina carpeta images de assets, todo archivo queda la raiz.Se borra importación de Footer en Layout por no uso
